Quick hack to build binary when installed via Quicklisp
When I originally wrote this, it was assumed that dependencies
would be configured directly via ASDF (the old school symlink
farm approach) which isn't very useful today when people are
most likely to install this and all dependencies via Quicklisp.

One approach to fixing this would be to introduce a separate
'configure' step that determines paths to all dependencies.
I'm not particularly motivated to implement this.

As a quick easy alternative, just remove "--no-userinit" when
building with SBCL. So long is Quicklisp is loaded in their
.sbclrc, this works just fine.

Fixes #2